# Optimize NEON encoding for high dynamic range blocks