We can improve the encoding performance on Apple silicon configurations with the following enhancements

- Replace the switch with a for loop
- Invert qlp_factors & remove else conditions as branches not taken are returned out

13% - 17% improvement depending on compiler, configuration and .wav

Clang 16 or GCC 13 necessary

| Platform | MBP18,2 | Mac14,12 | AMD Ryzen 9 7900X x 24 |
| Compiler | GCC 13 | Clang 16 | GCC 13 | Clang 16 | GNU 13.2.1 | 
| Original | 6.554 | 6.176 | 5.725 | 5.377 | 5.018 |
| Optimized | 5.642 | 5.131 | 4.803 | 4.672 | 5 |
| Overall | 14% | 17% | 16% | 13% | 0