@Developer-Ecosystem-Engineering 2024-03-11
Hi @ktmf01 

The optimization is designed to improve high bit rate audio. The test WAVs are detailed below. It's all stereo, 24bps, music files across many genres. 

This table was against macOS 14.4 + 16 inch MBP M2. 
| | Duration | Sample rate | Base time (s) | Optimized time (s) | Improvement (x times)
| ------------- | ------------- | ------------- | ------------- | ------------- | ------------- |
| Sample 1 | 08:47 | 44.1 | 1.843 | 1.685 | 1.09 |
| Sample 2 | 22:35 | 48 |  4.856 | 4.376 | 1.11 |
| Sample 3 | 15:36 | 88.2 | 8.28 | 6.537 | 1.27 |

> This change is not covered by fuzzing, provides a modest improvement for a very niche use. 

We'd be happy to review the PR further, including if you'd like additional fuzzing tests. A performance improvement for high res audio seems inline with FLAC use cases, i.e. https://github.com/xiph/flac/issues/669