@Developer-Ecosystem-Engineering 2024-03-14
Appreciate the dialog, thank you for taking the time to review and consider these changes!  

We made it exclusive because the benefit was much more pronounced on that architecture and flat on others. You are correct that it could be enabled for other architectures and that would also resolve the fuzzing issue. 

We'd be happy to modify the PR to not constrain it to only NEON. If there are specific WAVs and presets the project would like to see it tested against let us know, we will update with the results with those inputs and the expanded architecture.