@ktmf01 2024-05-16
> Okay, nevermind, this function is indeed used for 24-bit material.

While this was not the intent of this PR, thank you for (indirectly) pointing me to this. I've merged two PRs, #700 and #702, which should make `flac` use the function you optimized, `FLAC__lpc_compute_residual_from_qlp_coefficients_limit_residual`, way less often.

Could you perhaps rebase this PR against the current state of the repository, and re-evaluate this PR?